+++ spec/api/v1/short_url_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 describe API::V1::ShortUrl, type: :api do
-  let(:short_url) { create :short_url }
+  let!(:short_url) { create :short_url }
 
   it 'returns 404 when short_url is not found' do
     get 'api/v1/a_short_code'
@@ -24,6 +24,25 @@ describe API::V1::ShortUrl, type: :api do
 
     it 'returns 302' do
       expect(last_response.headers['Location']).to eq(short_url.url)
+    end
+  end
+
+  describe 'stats' do
+    let(:expected_response) do
+      {
+        'startDate' => short_url.start_date.iso8601,
+        'redirectCount' => short_url.redirect_count,
+        'lastSeenDate' => short_url.last_seen_date
+      }
+    end
+
+    before do
+      get "api/v1/#{short_url.short_code}/stats"
+    end
+
+    it 'returns the correct response' do
+      response = JSON.parse(last_response.body)
+      expect(response).to eq(expected_response)
     end
   end
 end